New redundant frameskip wrappers and no-op removal
No-op removal is simple, remove one action from action space and remap actions to align correctly with original action space

Surround is somewhat unique in that it's more like a turn-based game than other Atari games. Thus it's impoortant for RL transitions to actually fall all the in game transitions particularly for RL approaches like DQN otherwise it stores redundant transitions which seems to significantly hurt training efficacy.

Thus, use a redundant frameskip wrapper for this game to repeatedly step with actions until the state truly changes.

Interestingly, there's a "blip" of blue rectangle in the top left of the surround game which seems to flash whenever the environment is stepped. (This has at least been my observation in implementing and debugging this wrapper). In the future, it's possible to hardcode the wrapper to check for the blips and only step when the blip flashes as the game steps. This seems somewhat hardcoded but would be more efficient. Current implementation instead crops out this blip and does a np.allclose to compare consecutive frames and continues until env done or scene changes. The reason the blip needs to be cropped is that the frame with the blip (env just stepped) and frames after the blip before env step have no game state change, it's just that the first frame has the blip and the frames after the blip do not. Thus, we need to crop out the blip to compare frames correctly.