WIP rainbow dqn implementation
This implementation is already sufficient for actually beating the surround rule-based opponent as far as I can tell at around 20k episodes in the 1 round format.

This implementation has double dqn, PER, and dueling dqn already inherited from the previous agents/dqn.py and agents/dueling_dqn.py files.
The addition of the distribution learning for C51 is added with distributional loss and noisy MLP are added to this implementation.

What's missing from this implementation is multi-step learning. This is WIP but the current results for this env are already quite successful.

Hyperparameters are pretty much as in the original paper:
https://arxiv.org/pdf/1710.02298

There are some slight differences in the env setup, due to how the paper is training for all Atari games, and we make special env processing steps for surround.
Some of these includes only playing 1 round, using survival reward shaping, redudant frame skip instead of constant frame skip of 4, no frame stacking (as Atari Surround can be viewed as a turn-based game with no need to have temporal history (no velocity, etc)).